wip templatisation avec jinja

+++ src/qcmgen/template_mgr.py
@@ -1,6 +1,8 @@
+import re
 import importlib.resources
 from pathlib import Path
 from .common.utils import UserError
+from .common.jinja_templating import renderjinja
 from .amc_commands import prepare_project
 SUBFOLDERS = [
     "cr",
@@ -14,6 +16,25 @@ SUBFOLDERS = [
     "scans",
     "copies",
 ]
+
+
+def renderjinjaresources(sourcedir,outpath,data):
+    """extrait les fichiers des templates jinja localisés dans le reptoire ressources.
+
+    Args:
+        outpath (_type_): repertoire de destination des fichiers générés à partir des templates jinja
+        data (_type_): _description_
+    """
+    prj_folder = Path(outpath)
+    for fi in importlib.resources.files(sourcedir).iterdir():
+        if fi.is_file() and fi.name.endswith(".jinja"):
+            contenu=renderjinja("qcmgen","resources/amc_bootstrap_template",fi.name)
+            print(f"generation du fichier {fi.name}")
+            # l'extension .jinja est retirée du nom de fichier pour crérer le fichier de destinataiton
+            outfilename=re.sub(r'\.jinja$','',fi.name)
+            with (prj_folder/outfilename).open("w", encoding="utf-8") as outf:
+                outf.write(contenu)
+
 
 def get_curr_project_dir(folder:str="."):
     """retourne le repertoire du projet en validant qu'il s'agit bien d'un repertoire AMC
@@ -32,18 +53,20 @@ def copy_embedd_data(sourcedir,outpath):
     vers le répertoire outpath
     """
     for fi in importlib.resources.files(sourcedir).iterdir():
-        if fi.is_file():
+        if fi.is_file() and not fi.name.endswith(".jinja"):
             print(f"copie du fichier {fi.name}")
             with (fi.open("r", encoding="utf-8") as f):
                 contenu = f.read()
                 with (outpath/fi.name).open("w", encoding="utf-8") as outf:
                     outf.write(contenu)
 
-def copy_amcsources(outpath):
+def copy_amcsources(outpath,projectdata):
     """copie les fichiers contenus dans le repertoire qcmgen.resources.amc_bootstrap_template du package
     vers le répertoire outpath
     """
     copy_embedd_data("qcmgen.resources.amc_bootstrap_template",outpath)
+    renderjinjaresources("qcmgen.resources.amc_bootstrap_template",outpath,projectdata)
+
 
 
 def copy_questionbuilder_script(outpath):
@@ -63,7 +86,7 @@ def create_gitignore(outpath):
 !*.tex""")
 
 
-def create_project_structure(folder, overwrite):
+def create_project_structure(folder, overwrite,projectdata):
     """
     crée le repertoire AMC à partir de la structure de projet par défaut et 
     des fichiers templates contenus dans le repertoire qcmgen.resources.amc_bootstrap_template 
@@ -71,7 +94,7 @@ def create_project_structure(folder, overwrite):
     """
 
     def __addcontent(outpath):
-        copy_amcsources(prj_folder)
+        copy_amcsources(prj_folder,projectdata)
         copy_questionbuilder_script(prj_folder)
         create_gitignore(prj_folder)
         prepare_project(prj_folder)
